Refactor variable resolution into ReflectionVariableResolver
Extracted variable resolution logic from DefaultAiServices to a new utility class, ReflectionVariableResolver. This improves code modularity, reusability, and readability by centralizing the reflection-based variable handling logic.
